backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.10.0
pydantic-settings>=2.1.0

# AI
groq>=0.4.1

# Embeddings (lightweight, no torch dependency)
fastembed>=0.2.0

# Vector Database
qdrant-client>=1.7.0

# Database
psycopg2-binary>=2.9.0

# Utilities
python-dotenv==1.0.0
httpx>=0.24.0,<0.26.0

# Text Processing
beautifulsoup4==4.12.3
pypdf==4.0.1

# Email validation
email-validator>=2.0.0
